requirements.txt
crewai
crewai[tools]
crewai_tools
python-dotenv
streamlit
matplotlib
pandas
numpy
beautifulsoup4